refactor(#1390): fold the ten injected callbacks into one Deps struct
They are dependencies, not state: `init/1` reads them out of the resolved
plan and every later touch is a call. Grouping them takes Session.Server's
top-level state from 85 keys to 76 and stops EventRouter's state contract
carrying a bare closure among fields that genuinely are state.

`start_link/1`'s keyword API is unchanged — all ten option keys are
spelled exactly as before, so no producer of a session plan had to move.
That is what keeps this at six files rather than fifteen.

EventRouter matches a PLAIN map, not `%Deps{}`: Deps already names
`EventRouter.query_window_open?()` for a field type, so matching the
struct there would close a compile/export cycle. A struct is a map, so the
production value matches. Its absent arm keeps the `false` default and
deliberately does NOT fall back to `%Deps{}`, whose own default for that
field is the real DB read — handing it over would cost EventRouter its
"No Repo" purity in any test that omits `deps`.

Two of the twelve read sites were `Map.get(state, key, default)`, tolerant
of a process predating the field across a hot reload. Bundling hardened
them and the #581 hot-reload test caught it; they are now
`Map.get(state, :deps, %Deps{})`, exactly equivalent because the struct's
defaults for those two fields are the values those calls already passed.

The nine callback `@type`s move to `deps.ex` with their typedocs,
transplanted mechanically rather than retyped so the prose recording which
Boundary cycle each closure dodges travels with them. `restored_away` is
not one of the ten and stays put.

The EventRouter test builds a REAL `%Deps{}` through a small constructor
rather than nesting the override inside `base_state/1`: a fixture that
lies about the shape would keep 13 call sites expressing a shape
production no longer has.

Not measured here: Dialyzer, Credo, Sobelow and the full suite have not
been run — only `server_test.exs` + `event_router_test.exs`, 847 green. So
"behaviour-free" is established for those two files and asserted, not
proven, beyond them. The 85-to-76 count is arithmetic on the typedef, not
re-read off a live process.
